sha3: split simd/{arm64,avx2}.rs into load/store/wrappers submodules
Reorganize the NEON (arm64) and AVX2 SIMD backends from monolithic
files into per-concern submodules:

  simd/arm64.rs -> simd/arm64/{wrappers,load,store}.rs  (+ mod shim)
  simd/avx2.rs  -> simd/avx2/{wrappers,load,store}.rs   (+ mod shim)

- wrappers: math wrappers + the KeccakItem<N> impl (and the arm64
  `uint64x2_t` type alias, re-exported from the shim).
- load:     load_block, load_last, and the Absorb<N> impl.
- store:    store_block and the Squeeze{2,4} impl.

Pure code move: function bodies are unchanged. The SIMD backends are
excluded from the portable hax extraction, so the F* extraction is
byte-for-byte identical and the sha3-hax CI (extract/lax/type-check)
is unaffected.
